+++ web/src/api/APIClient.ts
@@ -1,4 +1,3 @@
-import {Action, DownloadClient, Filter, Indexer, Network} from "../domain/interfaces";
 import {baseUrl, sseBaseUrl} from "../utils";
 
 function baseClient(endpoint: string, method: string, { body, ...customConfig}: any = {}) {
