+++ web/src/components/inputs/select.tsx
@@ -2,7 +2,6 @@ import { Fragment } from "react";
 import { MultiSelect as RMSC}  from "react-multi-select-component";
 import { Transition, Listbox } from "@headlessui/react";
 import { CheckIcon, SelectorIcon } from '@heroicons/react/solid';
-import { Action, DownloadClient } from "../../domain/interfaces";
 import { classNames, COL_WIDTHS } from "../../utils";
 import { Field } from "formik";
