+++ web/src/types/Global.d.ts
@@ -1,0 +1,3 @@
+interface APP {
+    baseUrl: string;
+}
