+++ web/src/types/Release.d.ts
@@ -1,0 +1,36 @@
+interface Release {
+    id: number;
+    filter_status: string;
+    rejections: string[];
+    indexer: string;
+    filter: string;
+    protocol: string;
+    title: string;
+    size: number;
+    raw: string;
+    timestamp: Date
+    action_status: ReleaseActionStatus[]
+}
+
+interface ReleaseActionStatus {
+    id: number;
+    status: string;
+    action: string;
+    type: string;
+    rejections: string[];
+    timestamp: Date
+}
+
+interface ReleaseFindResponse {
+    data: Release[];
+    next_cursor: number;
+    count: number;
+}
+
+interface ReleaseStats {
+    total_count: number;
+    filtered_count: number;
+    filter_rejected_count: number;
+    push_approved_count: number;
+    push_rejected_count: number;
+}
